Move mock signer svc under test
